fix CPD-842: Overview Stats Optimization and Bug Fixes

+++ src/api/schemas/reports_schema.py
@@ -13,7 +13,7 @@ class SectorIndustryReportSchema(Schema):
     cycle_count = fields.Integer()
     emails_sent = fields.Integer()
     emails_clicked = fields.Integer()
-    emails_clicked_ratio = fields.Integer()
+    emails_clicked_ratio = fields.Float()
 
 
 class AggregateReportsSchema(Schema):
